% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ancestor}
\alias{ancestor}
\title{ancestor}
\usage{
ancestor(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
ancestor data
}
\description{
access ancestor data
}
\author{
guangchuang yu
}
