% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-treedata.R
\name{as.treedata.phylo}
\alias{as.treedata.phylo}
\title{as.treedata}
\usage{
\method{as.treedata}{phylo}(tree, boot = NULL, ...)
}
\arguments{
\item{tree}{input tree, a \code{phylo} object}

\item{boot}{optional, can be bootstrap value from ape::boot.phylo}

\item{...}{additional parameters}
}
\description{
convert phylo to treedata
}
\details{
converting phylo object to treedata object
}
\author{
Guangchuang Yu
}
