% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-ultrametric.R
\name{as.ultrametric}
\alias{as.ultrametric}
\title{as.ultrametric}
\usage{
as.ultrametric(tree, ...)
}
\arguments{
\item{tree}{tree object}

\item{...}{additional parameters}
}
\value{
treedata or phylo object
}
\description{
as.ultrametric
}
