% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast.R, R/mega.R
\name{read.beast}
\alias{read.beast}
\alias{read.mrbayes}
\alias{read.mega}
\title{read.beast}
\usage{
read.beast(file)

read.mrbayes(file)

read.mega(file)
}
\arguments{
\item{file}{beast file}
}
\value{
treedata object
}
\description{
read beast/mrbayes/mega Nexus output
}
\examples{
file <- system.file("extdata/BEAST", "beast_mcc.tree", package="treeio")
read.beast(file)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
