% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edgeNum2nodeNum}
\alias{edgeNum2nodeNum}
\title{edgeNum2nodeNum}
\usage{
edgeNum2nodeNum(jp, edgeNum)
}
\arguments{
\item{jp}{a \code{jplace} object}

\item{edgeNum}{the edge number of a placement, the \code{edge_num} column
of a jplace file. EPA and pplacer number the edges by a post-order
traversal, they are not the row numbers of \code{jp@phylo$edge}}
}
\value{
the node number of the reference tree, \code{NA} for an edge
number that is not in the tree
}
\description{
convert edge number to node number for EPA/pplacer output
}
\examples{
jpfile <- system.file("extdata", "sample.jplace", package="treeio")
jp <- read.jplace(jpfile)
edgeNum2nodeNum(jp, c(0, 1, 2))
}
\author{
Guangchuang Yu
}
