% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-find.spt.R
\name{find.spt}
\alias{find.spt}
\title{find.spt method}
\usage{
find.spt(x, from, to, ...)
}
\arguments{
\item{x}{a igraph object}

\item{from}{a specific node of network.}

\item{to}{other nodes of the network, length of it must 
be larger than 2.}

\item{...}{additional parameters}
}
\value{
phylo object
}
\description{
find.spt method
}
