% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\name{get.placements}
\alias{get.placements}
\alias{get.placements.jplace}
\title{get.placements}
\usage{
get.placements(tree, ...)

\method{get.placements}{jplace}(tree, by = "best", ...)
}
\arguments{
\item{tree}{tree object}

\item{...}{additional parameters}

\item{by}{one of 'best' and 'all'}
}
\value{
placement tibble
}
\description{
access placement information
}
\details{
The \code{node} column is the node number of the reference tree, that
is the node the branch leads to (a tip or an internal node, numbered
1:Ntip for the tips and Ntip+1:Ntip+Nnode for the internal nodes). It is
translated from the \code{edge_num} of the jplace file, which EPA and pplacer
number by a post-order traversal of the tree; \code{edgeNum2nodeNum()} does
that translation, #22
}
