% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-get-fields.R
\docType{methods}
\name{get.fields}
\alias{get.fields}
\alias{get.fields,treedata-method}
\title{get.fields method}
\usage{
get.fields(object, ...)

\S4method{get.fields}{treedata}(object)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast}, \code{hyphy}, \code{codeml}, \code{codeml_mlc}, \code{paml_rst} object}

\item{...}{additional parameter}
}
\value{
available annotation variables
}
\description{
get.fields method
}
