% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/codeml.R
\docType{methods}
\name{get.subs}
\alias{get.subs}
\alias{get.subs,codeml-method}
\title{get.subs method}
\usage{
get.subs(object, type, ...)

\S4method{get.subs}{codeml}(object, type, ...)
}
\arguments{
\item{object}{paml_rst object}

\item{type}{one of 'marginal_subs', 'marginal_AA_subs',
'joint_subs' or 'joint_AA_subs'.}

\item{...}{additional parameter}
}
\value{
data.frame
}
