% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/codeml.R
\docType{methods}
\name{get.tipseq}
\alias{get.tipseq}
\alias{get.tipseq,codeml-method}
\title{get.tipseq method}
\usage{
get.tipseq(object, ...)

\S4method{get.tipseq}{codeml}(object, ...)
}
\arguments{
\item{object}{one of paml_rst or codeml object}

\item{...}{additional parameter}
}
\value{
character
}
