% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-phylo.R
\name{get.tree}
\alias{get.tree}
\title{get.tree}
\usage{
get.tree(x, ...)
}
\arguments{
\item{x}{tree object}

\item{...}{additional parameters}
}
\value{
phylo object
}
\description{
access phylo slot
}
\author{
Guangchuang Yu
}
