% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{get.treeinfo}
\alias{get.treeinfo}
\title{get.treeinfo method}
\usage{
get.treeinfo(object, layout = "phylogram", ladderize = TRUE,
  right = FALSE, ...)
}
\arguments{
\item{object}{jplace object}

\item{layout}{layout}

\item{ladderize}{ladderize, logical}

\item{right}{logical, parameter for ladderize}

\item{...}{additional parameter}
}
\value{
data.frame
}

