% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-groupClade.R
\docType{methods}
\name{groupClade}
\alias{groupClade}
\alias{groupClade,beast-method}
\alias{groupClade,codeml-method}
\alias{groupClade,jplace-method}
\alias{groupClade,phangorn-method}
\alias{groupClade,phylip-method}
\alias{groupClade,phylo-method}
\alias{groupClade,treedata-method}
\title{groupClade method}
\usage{
groupClade(object, node, group_name = "group", ...)

\S4method{groupClade}{beast}(object, node, group_name = "group")

\S4method{groupClade}{codeml}(object, node, group_name = "group")

\S4method{groupClade}{jplace}(object, node, group_name = "group")

\S4method{groupClade}{treedata}(object, node, group_name = "group")

\S4method{groupClade}{phylip}(object, node, group_name = "group")

\S4method{groupClade}{phylo}(object, node, group_name = "group")

\S4method{groupClade}{phangorn}(object, node, group_name = "group")
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{node}{a internal node or a vector of internal nodes}

\item{group_name}{name of the group, 'group' by default}

\item{...}{additional parameter}
}
\value{
group index
}
\description{
group selected clade
}

