% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{groupClade}
\alias{groupClade}
\title{groupClade}
\usage{
groupClade(.data, .node, ...)
}
\arguments{
\item{.data}{tree object (phylo, treedata, tbl_tree, ggtree etc.)}

\item{.node}{selected nodes}

\item{...}{additional parameter}
}
\value{
updated tree with group information or group index
}
\description{
grouping clades
}
\author{
Guangchuang Yu
}
