% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-groupOTU.R
\docType{methods}
\name{groupOTU}
\alias{groupOTU}
\alias{groupOTU,treedata-method}
\alias{groupOTU,phylo-method}
\title{groupOTU method}
\usage{
groupOTU(object, focus, group_name = "group", ...)

\S4method{groupOTU}{treedata}(object, focus, group_name = "group", ...)

\S4method{groupOTU}{phylo}(object, focus, group_name = "group", ...)
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{focus}{a vector of tip (label or number) or a list of tips.}

\item{group_name}{name of the group, 'group' by default}

\item{...}{additional parameter}
}
\value{
group index
}
\description{
group tree based on selected OTU, will traceback to MRCA
}
