% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{groupOTU}
\alias{groupOTU}
\title{groupOTU}
\usage{
groupOTU(.data, .node, ...)
}
\arguments{
\item{.data}{tree object (phylo, treedata, tbl_tree, ggtree etc.)}

\item{.node}{selected nodes}

\item{...}{additional parameter}
}
\value{
updated tree with group information or group index
}
\description{
grouping OTUs
}
\author{
guangchuang yu
}
