% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/isTip.R
\name{isTip}
\alias{isTip}
\alias{isTip.tbl_tree}
\alias{isTip.phylo}
\alias{isTip.treedata}
\title{isTip}
\usage{
isTip(.data, .node, ...)

\method{isTip}{tbl_tree}(.data, .node, ...)

\method{isTip}{phylo}(.data, .node, ...)

\method{isTip}{treedata}(.data, .node, ...)
}
\arguments{
\item{.data}{phylo, treedata or tbl_tree object}

\item{.node}{node number}

\item{...}{additional parameters}
}
\value{
logical value
}
\description{
whether the node is a tip
}
\author{
Guangchuang Yu
}
