% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{jplace-class}
\alias{jplace-class}
\title{Class "jplace"
This class stores phylogenetic placements}
\description{
Class "jplace"
This class stores phylogenetic placements
}
\section{Slots}{

\describe{
\item{\code{phylo}}{phylo object for tree structure}

\item{\code{treetext}}{newick tree string}

\item{\code{data}}{associated data}

\item{\code{extraInfo}}{extra information, reserve for merge_tree}

\item{\code{file}}{tree file}

\item{\code{placements}}{reserve for jplace file to store placement information}

\item{\code{info}}{extra information, e.g. metadata, software version etc.}
}}

\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}
