% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paml.R
\name{label_branch_paml}
\alias{label_branch_paml}
\title{label_branch_paml}
\usage{
label_branch_paml(tree, node, label)
}
\arguments{
\item{tree}{phylo object}

\item{node}{node number}

\item{label}{label of branch, e.g. #1}
}
\value{
updated phylo object
}
\description{
label branch for PAML to infer selection pressure using branch model
}
\author{
guangchuang yu
}

