% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{Nnode}
\alias{Nnode}
\title{Nnode}
\usage{
Nnode(tree, internal.only = TRUE)
}
\arguments{
\item{tree}{tree object}

\item{internal.only}{whether only count internal nodes}
}
\value{
number of nodes
}
\description{
number of nodes
}
\examples{
Nnode(rtree(30))
}
\author{
guangchuang yu
}

