% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\name{Nnode.treedata}
\alias{Nnode.treedata}
\title{Nnode}
\usage{
\method{Nnode}{treedata}(phy, internal.only = TRUE, ...)
}
\arguments{
\item{phy}{treedata object}

\item{internal.only}{whether only count internal nodes}

\item{...}{additional parameters}
}
\value{
number of nodes
}
\description{
number of nodes
}
\examples{
Nnode(rtree(30))
}
\author{
Guangchuang Yu
}
