% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{Ntip}
\alias{Ntip}
\title{Ntip}
\usage{
Ntip(tree)
}
\arguments{
\item{tree}{tree object}
}
\value{
number of tips
}
\description{
number of tips
}
\examples{
Ntip(rtree(30))
}
\author{
guangchuang yu

guangchuang yu
}

