% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{offspring}
\alias{offspring}
\title{offspring}
\usage{
offspring(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
offspring data
}
\description{
access offspring data
}
\author{
guangchuang yu
}
