% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{parent}
\alias{parent}
\title{parent}
\usage{
parent(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
parent data
}
\description{
access parent data
}
\author{
guangchuang yu
}
