% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.treedataList}
\alias{print.treedataList}
\title{print}
\usage{
\method{print}{treedataList}(x, ...)
}
\arguments{
\item{x}{a list of treedata objects}

\item{...}{no used}
}
\value{
message
}
\description{
print information of a list of treedata objects
}
