% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtree.R
\name{read.jtree}
\alias{read.jtree}
\title{read.jtree}
\usage{
read.jtree(file)
}
\arguments{
\item{file}{tree file}
}
\value{
treedata object
}
\description{
Import tree data from jtree file, which is JSON-based text and probably output by write.jtree
}
\author{
guangchuang yu
}
