% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeml.R
\name{read.codeml}
\alias{read.codeml}
\title{read.codeml}
\usage{
read.codeml(rstfile, mlcfile, tree = "mlc", type = "Joint")
}
\arguments{
\item{rstfile}{rst file}

\item{mlcfile}{mlc file}

\item{tree}{one of 'mlc' or 'rst'}

\item{type}{one of 'Marginal' or 'Joint'}
}
\value{
A \code{treedata} object
}
\description{
read baseml output
}
\examples{
rstfile <- system.file("extdata/PAML_Codeml", "rst", package="treeio")
mlcfile <- system.file("extdata/PAML_Codeml", "mlc", package="treeio")
read.codeml(rstfile, mlcfile)
}
\author{
Guangchuang Yu
}
