% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-utilities.R
\name{read.fasta}
\alias{read.fasta}
\title{read.fasta}
\usage{
read.fasta(fasta)
}
\arguments{
\item{fasta}{fasta file}
}
\value{
DNAbin or AAbin object
}
\description{
read FASTA file
}
\details{
This function supports both DNA or AA sequences
}
\author{
guangchuang yu
}
