% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphy.R
\name{read.hyphy.seq}
\alias{read.hyphy.seq}
\title{read.hyphy.seq}
\usage{
read.hyphy.seq(file)
}
\arguments{
\item{file}{output of hyphy ancestral sequence inference; nexus format}
}
\value{
DNAbin object
}
\description{
parse sequences from hyphy output
}
\examples{
ancseq <- system.file("extdata/HYPHY", "ancseq.nex", package="treeio")
read.hyphy.seq(ancseq)
}
\author{
Guangchuang Yu
}
