% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCTree.R
\name{read.mcmctree}
\alias{read.mcmctree}
\title{read.mcmctree}
\usage{
read.mcmctree(file, force.ultrametric = FALSE)
}
\arguments{
\item{file}{the output tree file of MCMCTree}

\item{force.ultrametric}{logical whether convert the tree 
to be ultrametric, if it is not ultrametric, default is FALSE.
When the tree is ultrametric, branch times will be calculated 
automatically.}
}
\value{
treedata object
}
\description{
read MCMCTree output Tree
}
\details{
The node age is not reported by MCMCTree, it is derived from the
branch lengths and stored in the \code{reltime} column. The credibility interval
of a node age, \code{[&95\%={lower, upper}]} in the file, is stored in the
\code{reltime_0.95_CI} column and can be plotted with
\code{geom_range(range='reltime_0.95_CI', center='reltime')}.
}
\examples{
file <- system.file("extdata/MCMCTree", "mcmctree_output.tree", package="treeio")
tr <- read.mcmctree(file)
tr
}
