% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega.R
\name{read.mega_tabular}
\alias{read.mega_tabular}
\title{read.mega_tabular}
\usage{
read.mega_tabular(file)
}
\arguments{
\item{file}{MEGA tabular file}
}
\value{
treedata object
}
\description{
parse tabular output of MEGA
}
\author{
Guangchuang Yu
}
