% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newick.R
\name{read.newick}
\alias{read.newick}
\title{read.newick}
\usage{
read.newick(file, node.label = "label", ...)
}
\arguments{
\item{file}{newick file}

\item{node.label}{parse node label as 'label' or 'support' value}

\item{...}{additional parameter, passed to 'read.tree'}
}
\value{
phylo or treedata object
}
\description{
read newick tree
}
\author{
guangchuang yu
}

