% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhx.R
\name{read.nhx}
\alias{read.nhx}
\title{read.nhx}
\usage{
read.nhx(file)
}
\arguments{
\item{file}{nhx file}
}
\value{
nhx object
}
\description{
read nhx tree file
}
\examples{
nhxfile <- system.file("extdata/NHX", "ADH.nhx", package="treeio")
read.nhx(nhxfile)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}

