% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paml_rst.R
\name{read.paml_rst}
\alias{read.paml_rst}
\title{read.paml_rst}
\usage{
read.paml_rst(rstfile, type = "Joint")
}
\arguments{
\item{rstfile}{rst file}

\item{type}{one of 'Marginal' or 'Joint'}
}
\value{
A \code{treedata} object
}
\description{
read rst file from paml (both baseml and codeml) output
}
\examples{
rstfile <- system.file("extdata/PAML_Baseml", "rst", package="treeio")
read.paml_rst(rstfile)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
