% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylip.R
\name{read.phylip.seq}
\alias{read.phylip.seq}
\title{read.phylip.seq}
\usage{
read.phylip.seq(file)
}
\arguments{
\item{file}{phylip file, currently only sequential format is supported}
}
\value{
DNAbin object
}
\description{
read aligned sequences from phylip format
}
\author{
guangchuang yu
}
