% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylip.R
\name{read.phylip.tree}
\alias{read.phylip.tree}
\title{read.phylip.tree}
\usage{
read.phylip.tree(file)
}
\arguments{
\item{file}{phylip file}
}
\value{
phylo or multiPhylo object
}
\description{
parse tree from phylip file
}
\author{
guangchuang yu
}
