% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloT.R
\name{read.phyloT}
\alias{read.phyloT}
\title{read.phyloT}
\usage{
read.phyloT(file, ...)
}
\arguments{
\item{file}{newick tree file}

\item{...}{additional parameters to read.tree}
}
\value{
phylo object
}
\description{
parse output from phyloT
}
\references{
\url{http://phylot.biobyte.de/}
}
\author{
guangchuang yu
}
