% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloxml.R
\name{read.phyloxml}
\alias{read.phyloxml}
\title{read.phyloxml}
\usage{
read.phyloxml(file)
}
\arguments{
\item{file}{phyloxml file}
}
\value{
treedata class
}
\description{
read.phyloxml
}
\examples{
xmlfile <- system.file("extdata/phyloxml", "test_x2.xml", package="treeio")
px <- read.phyloxml(xmlfile)
px
}
