% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ape.R, R/method-as-phylo.R,
%   R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{treedata}
\alias{read.tree}
\alias{read.nexus}
\alias{rtree}
\alias{write.tree}
\alias{write.nexus}
\alias{Nnode}
\alias{Ntip}
\alias{is.rooted}
\alias{root}
\alias{as.phylo}
\alias{\%>\%}
\alias{\%<>\%}
\alias{get.fields}
\alias{get.data}
\alias{as.treedata}
\alias{ancestor}
\alias{parent}
\alias{child}
\alias{offspring}
\alias{rootnode}
\alias{nodeid}
\alias{nodelab}
\alias{MRCA}
\alias{full_join}
\alias{inner_join}
\alias{as_tibble}
\alias{tibble}
\alias{.data}
\alias{drop.tip}
\alias{isTip}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape:as.phylo]{as.phylo()}}, \code{\link[ape:is.rooted]{is.rooted()}}, \code{\link[ape:Nnode]{Nnode()}}, \code{\link[ape:Ntip]{Ntip()}}, \code{\link[ape:read.nexus]{read.nexus()}}, \code{\link[ape:read.tree]{read.tree()}}, \code{\link[ape:root]{root()}}, \code{\link[ape:rtree]{rtree()}}, \code{\link[ape:write.nexus]{write.nexus()}}, \code{\link[ape:write.tree]{write.tree()}}}

  \item{dplyr}{\code{\link[dplyr:full_join]{full_join()}}, \code{\link[dplyr:inner_join]{inner_join()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:.data]{.data}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}, \code{\link[tibble:tibble]{tibble()}}}

  \item{tidytree}{\code{\link[tidytree:ancestor]{ancestor()}}, \code{\link[tidytree:as.phylo]{as.phylo}}, \code{\link[tidytree:as.treedata]{as.treedata()}}, \code{\link[tidytree:child]{child()}}, \code{\link[tidytree:drop.tip]{drop.tip()}}, \code{\link[tidytree:get.data]{get.data()}}, \code{\link[tidytree:get.fields]{get.fields()}}, \code{\link[tidytree:isTip]{isTip()}}, \code{\link[tidytree:MRCA]{MRCA()}}, \code{\link[tidytree:nodeid]{nodeid()}}, \code{\link[tidytree:nodelab]{nodelab()}}, \code{\link[tidytree:offspring]{offspring()}}, \code{\link[tidytree:parent]{parent()}}, \code{\link[tidytree:rootnode]{rootnode()}}, \code{\link[tidytree:treedata]{treedata()}}}
}}

