% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-reroot.R
\name{reroot}
\alias{reroot}
\alias{reroot.phylo}
\alias{reroot.treedata}
\title{reroot}
\usage{
reroot(object, node, ...)

\method{reroot}{phylo}(object, node, ...)

\method{reroot}{treedata}(object, node, ...)
}
\arguments{
\item{object}{tree object}

\item{node}{node to reroot}

\item{...}{additional parameters}
}
\value{
rerooted tree
}
\description{
re-root a tree
}
\author{
Guangchuang Yu
}
