% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-reroot.R
\name{root.phylo}
\alias{root.phylo}
\alias{root.treedata}
\title{root}
\usage{
\method{root}{phylo}(object, node, ...)

\method{root}{treedata}(object, node, ...)
}
\arguments{
\item{object}{tree object}

\item{node}{node to reroot}

\item{...}{additional parameters}
}
\value{
rerooted tree
}
\description{
re-root a tree
}
\author{
Guangchuang Yu
}
