% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rootnode}
\alias{rootnode}
\title{rootnode}
\usage{
rootnode(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
root node data
}
\description{
access root node data
}
\author{
guangchuang yu
}
