% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{\code{treedata} object}
}
\value{
print info
}
\description{
show method for \code{treedata} instance
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="treeio")
jp <- read.jplace(jp)
show(jp)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
