% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{taxa_rename}
\alias{taxa_rename}
\title{taxa_rename}
\usage{
taxa_rename(tree, name)
}
\arguments{
\item{tree}{tree object}

\item{name}{a two column data.frame contains original name in 1st column and new name in 2nd column}
}
\value{
updated tree object with new taxa name
}
\description{
rename taxa
}
\author{
guangchuang yu
}

