% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetime.R
\name{read.treetime}
\alias{read.treetime}
\alias{read.timetree}
\title{read.timetree}
\usage{
read.treetime(file)

read.timetree(file)
}
\arguments{
\item{file}{the output tree file of timetree}
}
\value{
treedata object
}
\description{
read timetree output
}
