% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtree.R
\name{write.jtree}
\alias{write.jtree}
\title{write.jtree}
\usage{
write.jtree(treedata, file = "")
}
\arguments{
\item{treedata}{\code{treedata} object}

\item{file}{output file. If file = "", print the output content on screen}
}
\value{
output file or file content on screen
}
\description{
Export \code{treedata} object to json tree file
}
\author{
guangchuang yu
}
