% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-jplace.R
\name{write.jplace}
\alias{write.jplace}
\title{write.jplace}
\usage{
write.jplace(x, outfile)
}
\arguments{
\item{x}{a jplace object.}

\item{outfile}{the output file name}
}
\description{
Export \code{jplace} object to jplace file.
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="treeio")
tr1 <- read.jplace(jp)
outfile <- tempfile()
write.jplace(tr1, outfile)
tr2 <- read.jplace(outfile)
tr2
}
