% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-beast.R
\name{write_beast_newick}
\alias{write_beast_newick}
\title{write.beast}
\usage{
write_beast_newick(
  treedata,
  file = "",
  append = FALSE,
  digits = 10,
  tree.prefix = ""
)
}
\arguments{
\item{treedata}{\code{treedata} object}

\item{file}{output file. If file = "", print the output content on screen}

\item{translate}{whether translate taxa labels}

\item{tree.name}{name of the tree}
}
\value{
output file or file content on screen
}
\description{
Export \code{treedata} object to BEAST Newick file. This is useful for making BEAST starting trees with metadata
}
\examples{
nhxfile <- system.file("extdata/NHX", "phyldog.nhx", package="treeio")
nhx <- read.nhx(nhxfile)
write_beast_newick(nhx)
}
\author{
Guangchuang Yu
}
