% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-treedata.R, R/phangorn.R
\name{as.treedata.phylo}
\alias{as.treedata.phylo}
\alias{as.treedata.pml}
\title{as.treedata}
\usage{
\method{as.treedata}{phylo}(tree, boot = NULL, ...)

\method{as.treedata}{pml}(tree, type = "ml", ...)
}
\arguments{
\item{tree}{input tree, a \code{phylo} object}

\item{boot}{optional, can be bootstrap value from ape::boot.phylo}

\item{...}{additional parameters}

\item{type}{one of 'ml' and 'bayes' for inferring ancestral sequences}
}
\description{
convert phylo to treedata
}
\details{
converting phylo object to treedata object
}
\author{
Guangchuang Yu

Yu Guangchuang
}
