% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-as-treedata.R
\name{as.treedata}
\alias{as.treedata}
\alias{as.treedata.phylo}
\title{as.treedata}
\usage{
as.treedata(tree, ...)

\method{as.treedata}{phylo}(tree, boot = NULL, ...)
}
\arguments{
\item{tree}{tree object}

\item{...}{additional parameter}

\item{boot}{optional, can be bootstrap value from ape::boot.phylo}
}
\value{
treedata object
}
\description{
converting phylo object to treedata object
}
\author{
guangchuang yu
}

