% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{child}
\alias{child}
\title{child}
\usage{
child(.data, ...)
}
\arguments{
\item{.data}{phylo or tbl_tree object}

\item{...}{additional parameters}
}
\value{
child data
}
\description{
access child data
}
\author{
guangchuang yu
}
