% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\name{get.placements}
\alias{get.placements}
\alias{get.placements.jplace}
\title{get.placements}
\usage{
get.placements(tree, ...)

\method{get.placements}{jplace}(tree, by = "best", ...)
}
\arguments{
\item{tree}{tree object}

\item{...}{additional parameters}

\item{by}{one of 'best' and 'all'}
}
\value{
placement tibble
}
\description{
access placement information
}
