% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-get-treetext.R
\docType{methods}
\name{get.treetext}
\alias{get.treetext}
\alias{get.treetext,treedata-method}
\title{get.treetext method}
\usage{
get.treetext(object, ...)

\S4method{get.treetext}{treedata}(object)
}
\arguments{
\item{object}{treedata object}

\item{...}{additional parameter}
}
\value{
phylo object
}
\description{
access tree text (newick text) from tree object
}
