% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-utilities.R
\name{read.fasta}
\alias{read.fasta}
\title{read.fasta}
\usage{
read.fasta(fasta, type = "auto")
}
\arguments{
\item{fasta}{fasta file}

\item{type}{sequence type of the input file, one of 'NT' or 'AA'.
Default is 'auto' and guess the sequence type automatically}
}
\value{
DNAbin or AAbin object
}
\description{
read FASTA file
}
\details{
This function supports both DNA or AA sequences
}
\author{
Guangchuang Yu
}
