% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtree.R
\name{read.iqtree}
\alias{read.iqtree}
\title{read.iqtree}
\usage{
read.iqtree(file)
}
\arguments{
\item{file}{IQ-TREE Newick text}
}
\value{
treedata object
}
\description{
parse IQ-TREE output
}
\author{
Guangchuang Yu
}
