% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCTree.R
\name{read.mcmctree}
\alias{read.mcmctree}
\title{read.mcmctree}
\usage{
read.mcmctree(file)
}
\arguments{
\item{file}{the output tree file of MCMCTree}
}
\value{
treedata object
}
\description{
read MCMCTree output Tree
}
\examples{
file <- system.file("extdata/MCMCTree", "mcmctree_output.tree", package="treeio")
tr <- read.mcmctree(file)
tr
}
