% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCTree.R
\name{read.mcmctree}
\alias{read.mcmctree}
\title{read.mcmctree}
\usage{
read.mcmctree(file, as.ultrametric = FALSE)
}
\arguments{
\item{file}{the output tree file of MCMCTree}

\item{as.ultrametric}{logical whether convert the tree 
to be ultrametric, if it is not ultrametric, default is FALSE.
When the tree is ultrametric, branch times will be calculated 
automatically.}
}
\value{
treedata object
}
\description{
read MCMCTree output Tree
}
\examples{
file <- system.file("extdata/MCMCTree", "mcmctree_output.tree", package="treeio")
tr <- read.mcmctree(file)
tr
}
