% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloxml.R
\name{read.phyloxml}
\alias{read.phyloxml}
\title{read.phyloxml}
\usage{
read.phyloxml(file)
}
\arguments{
\item{file}{phyloxml file}
}
\value{
treedata class or treedataList class
}
\description{
read.phyloxml
}
\examples{
xmlfile1 <- system.file("extdata/phyloxml", "test_x2.xml", package="treeio")
px1 <- read.phyloxml(xmlfile1)
px1
xmlfile2 <- system.file("extdata/phyloxml", "phyloxml_examples.xml", package="treeio")
px2 <- read.phyloxml(xmlfile2)
px2
}
