% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAxML.R
\name{read.raxml}
\alias{read.raxml}
\title{read.raxml}
\usage{
read.raxml(file)
}
\arguments{
\item{file}{RAxML bootstrapping analysis output}
}
\value{
raxml object
}
\description{
parse RAxML bootstrapping analysis output
}
\examples{
raxml_file <- system.file("extdata/RAxML", "RAxML_bipartitionsBranchLabels.H3", package="treeio")
read.raxml(raxml_file)
}
\author{
Guangchuang Yu
}

