% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R, R/method-as-phylo.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read.tree}
\alias{reexports}
\alias{read.nexus}
\alias{reexports}
\alias{rtree}
\alias{reexports}
\alias{write.tree}
\alias{reexports}
\alias{Nnode}
\alias{reexports}
\alias{Ntip}
\alias{reexports}
\alias{as.phylo}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{get.fields}
\alias{reexports}
\alias{as.treedata}
\alias{reexports}
\alias{full_join}
\alias{reexports}
\alias{data_frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}}, \code{\link[ape]{rtree}}, \code{\link[ape]{write.tree}}, \code{\link[ape]{Nnode}}, \code{\link[ape]{Ntip}}, \code{\link[ape]{as.phylo}}}

  \item{dplyr}{\code{\link[dplyr]{full_join}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{data_frame}}}

  \item{tidytree}{\code{\link[tidytree]{get.fields}}, \code{\link[tidytree]{as.treedata}}}
}}

