% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ape.R, R/method-as-phylo.R,
%   R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{treedata}
\alias{read.tree}
\alias{read.nexus}
\alias{rtree}
\alias{write.tree}
\alias{Nnode}
\alias{Ntip}
\alias{is.rooted}
\alias{as.phylo}
\alias{\%>\%}
\alias{get.fields}
\alias{get.data}
\alias{as.treedata}
\alias{ancestor}
\alias{parent}
\alias{child}
\alias{offspring}
\alias{rootnode}
\alias{full_join}
\alias{data_frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}}, \code{\link[ape]{rtree}}, \code{\link[ape]{write.tree}}, \code{\link[ape]{Nnode}}, \code{\link[ape]{Ntip}}, \code{\link[ape]{is.rooted}}, \code{\link[ape]{as.phylo}}}

  \item{dplyr}{\code{\link[dplyr]{full_join}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{data_frame}}}

  \item{tidytree}{\code{\link[tidytree]{treedata}}, \code{\link[tidytree]{get.fields}}, \code{\link[tidytree]{get.data}}, \code{\link[tidytree]{as.treedata}}, \code{\link[tidytree]{ancestor}}, \code{\link[tidytree]{parent}}, \code{\link[tidytree]{child}}, \code{\link[tidytree]{offspring}}, \code{\link[tidytree]{rootnode}}}
}}

