% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ape.R, R/method-as-phylo.R,
%   R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{treedata}
\alias{read.tree}
\alias{read.nexus}
\alias{rtree}
\alias{write.tree}
\alias{write.nexus}
\alias{Nnode}
\alias{Ntip}
\alias{is.rooted}
\alias{root}
\alias{as.phylo}
\alias{\%>\%}
\alias{get.fields}
\alias{get.data}
\alias{as.treedata}
\alias{ancestor}
\alias{parent}
\alias{child}
\alias{offspring}
\alias{rootnode}
\alias{nodeid}
\alias{nodelab}
\alias{MRCA}
\alias{full_join}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}, \code{\link[ape:root]{is.rooted}}, \code{\link[ape:summary.phylo]{Nnode}}, \code{\link[ape:summary.phylo]{Ntip}}, \code{\link[ape]{read.nexus}}, \code{\link[ape]{read.tree}}, \code{\link[ape]{root}}, \code{\link[ape]{rtree}}, \code{\link[ape]{write.nexus}}, \code{\link[ape]{write.tree}}}

  \item{dplyr}{\code{\link[dplyr:mutate-joins]{full_join}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidytree}{\code{\link[tidytree]{ancestor}}, \code{\link[tidytree]{as.treedata}}, \code{\link[tidytree]{child}}, \code{\link[tidytree:get.data-methods]{get.data}}, \code{\link[tidytree:get.fields-methods]{get.fields}}, \code{\link[tidytree]{MRCA}}, \code{\link[tidytree]{nodeid}}, \code{\link[tidytree]{nodelab}}, \code{\link[tidytree]{offspring}}, \code{\link[tidytree]{parent}}, \code{\link[tidytree]{rootnode}}, \code{\link[tidytree]{treedata}}}
}}

