% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-phylo.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{as.phylo}
\alias{read.tree}
\alias{reexports}
\alias{rtree}
\alias{write.tree}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}, \code{\link[ape]{read.tree}}, \code{\link[ape]{rtree}}, \code{\link[ape]{write.tree}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\keyword{internal}

