% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_taxa}
\alias{rename_taxa}
\title{rename_taxa}
\usage{
rename_taxa(tree, data, key, value)
}
\arguments{
\item{tree}{tree object, either treedata or phylo}

\item{data}{data frame}

\item{key}{column in data that match tip label}

\item{value}{column in data for rename tip label}
}
\value{
tree object
}
\description{
rename tip label of phylogenetic tree
}
\author{
Guangchuang Yu
}
