% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-reroot.R
\name{root.phylo}
\alias{root.phylo}
\alias{root.treedata}
\title{root}
\usage{
\method{root}{phylo}(object, outgroup, node = NULL,
  resolve.root = TRUE, ...)

\method{root}{treedata}(object, outgroup, node = NULL,
  resolve.root = TRUE, ...)
}
\arguments{
\item{object}{tree object}

\item{outgroup}{a vector of mode numeric or character specifying the new outgroup}

\item{node}{node to reroot}

\item{resolve.root}{a logical specifying whether to resolve the new root as a bifurcating node}

\item{...}{additional parameters passed to ape::root.phylo}
}
\value{
rerooted tree
}
\description{
re-root a tree
}
\author{
Guangchuang Yu
}
