% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-reroot.R
\name{root.treedata}
\alias{root.treedata}
\title{root}
\usage{
\method{root}{treedata}(phy, outgroup, node = NULL, edgelabel = TRUE, ...)
}
\arguments{
\item{phy}{tree object}

\item{outgroup}{a vector of mode numeric or character specifying the new outgroup}

\item{node}{node to reroot}

\item{edgelabel}{a logical value specifying whether to treat node labels as
edge labels and thus eventually switching them so that they are associated
with the correct edges.}

\item{...}{additional parameters passed to ape::root.phylo}
}
\value{
rerooted treedata
}
\description{
re-root a tree
}
