% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\docType{methods}
\name{show,beast-method}
\alias{show}
\alias{show,beast-method}
\alias{show,codeml-method}
\alias{show,codeml_mlc-method}
\alias{show,hyphy-method}
\alias{show,paml_rst-method}
\alias{show,phylip-method}
\alias{show,r8s-method}
\alias{show,treedata-method}
\title{show method}
\usage{
\S4method{show}{beast}(object)

\S4method{show}{codeml}(object)

\S4method{show}{codeml_mlc}(object)

show(object)

\S4method{show}{phylip}(object)

\S4method{show}{paml_rst}(object)

\S4method{show}{r8s}(object)

\S4method{show}{hyphy}(object)

\S4method{show}{treedata}(object)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}
}
\value{
print info
}
\description{
show method for \code{jplace} instance
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="treeio")
jp <- read.jplace(jp)
show(jp)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}

