% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sibling}
\alias{sibling}
\title{sibling}
\usage{
sibling(.data, ...)
}
\arguments{
\item{.data}{phylo or tbl_tree object}

\item{...}{additional parameters}
}
\value{
sibling
}
\description{
access sibling data
}
\author{
guangchuang yu
}
