% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{treedata-class}
\alias{treedata-class}
\title{Class "treedata"
This class stores phylogenetic tree with associated data}
\description{
Class "treedata"
This class stores phylogenetic tree with associated data
}
\section{Slots}{

\describe{
\item{\code{phylo}}{phylo object for tree structure}

\item{\code{treetext}}{newick tree string}

\item{\code{data}}{associated data}

\item{\code{extraInfo}}{extra information, reserve for merge_tree}

\item{\code{file}}{tree file}
}}
\author{
guangchuang yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}

