% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio-package.R
\docType{package}
\name{treeio-package}
\alias{treeio}
\alias{treeio-package}
\title{treeio: Base Classes and Functions for Phylogenetic Tree Input and Output}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'treeio' is an R package to make it easier to import and store phylogenetic tree with associated data; and to link external data from different sources to phylogeny. It also supports exporting phylogenetic tree with heterogeneous associated data to a single tree file and can be served as a platform for merging tree with associated data and converting file formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/treeio (devel)}
  \item \url{https://docs.ropensci.org/treeio/ (docs)}
  \item \url{https://www.amazon.com/Integration-Manipulation-Visualization-Phylogenetic-Computational-ebook/dp/B0B5NLZR1Z/ (book)}
  \item \doi{10.1093/molbev/msz240 (paper)}
  \item Report bugs at \url{https://github.com/YuLab-SMU/treeio/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Tommy Tsan-Yuk Lam \email{tylam.tommy@gmail.com} [contributor, thesis advisor]
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Bradley Jones \email{brj1@sfu.ca} [contributor]
  \item Casey Dunn \email{casey_dunn@brown.edu} [contributor]
  \item Tyler Bradley \email{tcb85@drexel.edu} [contributor]
  \item Konstantinos Geles \email{konstantinos.geles@studenti.unicz.it} [contributor]
}

}
\keyword{internal}
