% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-beast.R
\name{write.beast}
\alias{write.beast}
\title{write.beast}
\usage{
write.beast(treedata, file = "", translate = TRUE,
  tree.name = "UNTITLED")
}
\arguments{
\item{treedata}{\code{treedata} object}

\item{file}{output file. If file = "", print the output content on screen}

\item{translate}{whether translate taxa labels}

\item{tree.name}{name of the tree}
}
\value{
output file or file content on screen
}
\description{
Export \code{treedata} object to BEAST NEXUS file. This function was adopted and modified from ape::write.nexus
}
\examples{
nhxfile <- system.file("extdata/NHX", "phyldog.nhx", package="treeio")
nhx <- read.nhx(nhxfile)
write.beast(nhx)
}
\author{
guangchuang yu
}
